requirements.txt
aiogram==3.10.0
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
aiohttp==3.9.1